% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAfisher_default.R
\name{GSAfisher.default}
\alias{GSAfisher.default}
\title{A GSAfisher default method}
\usage{
\method{GSAfisher}{default}(obj, ...)
}
\arguments{
\item{obj}{The object you want to calculate the Fischer value}
}
\description{
This function calculates the GSAfisher value for each pvalue
}
\examples{
p.val <- runif(1000,0,0.25)
class(p.val) <- "GSAfisher"
GSAfisher(p.val)
}
\keyword{GSAfisher}
