% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAfisher.R
\name{GSAfisher}
\alias{GSAfisher}
\title{A GSAfisher Function}
\usage{
GSAfisher(obj, ...)
}
\arguments{
\item{obj}{The object you want to calculate the Fischer value}
}
\description{
This function acts as a foundation to the other methods
}
\examples{
GSAfisher()
}
\keyword{GSAfisher}
