% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAfisher_print.R
\name{print.GSAfisher}
\alias{print.GSAfisher}
\title{A print method for the class GSAfisher}
\usage{
\method{print}{GSAfisher}(obj, ...)
}
\arguments{
\item{obj}{The object you want to calculate the GSAfisher estatistic value}
}
\description{
This prints the GSAfisher estatistic value
}
\examples{
p.val <- runif(1000,0,0.25)
class(p.val) <- "GSAfisher"
print(p.val)
}
\keyword{GSAfisher}
