% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAfisher_summary.R
\name{summary.GSAfisher}
\alias{summary.GSAfisher}
\title{A summary method for the class GSAfisher}
\usage{
\method{summary}{GSAfisher}(obj, ...)
}
\arguments{
\item{obj}{The object you want to see this brief summary}
}
\description{
This prints a brief summary for the object of class GSAfisher
}
\examples{
p.val <- runif(1000,0,0.25)
class(p.val) <- "GSAfisher"
summary(p.val)
}
\keyword{GSAfisher}
